pyproject.toml
[project]
name = "map-asci"
version = "0.1.0"
description = "ASCII world map generator from a real land mask"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.1.0",
    "pillow>=11.0.0",
    "regionmask>=0.13.0",
    "xarray>=2024.10.0",
]

[project.scripts]
map-asci = "map_asci.cli:main"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/map_asci"]
